🍒 Initialize slt-acorn-plugin.
This is a base commit for an independent repo for the slt acorn plugin.

Keeping with the unix philosphy of doing one thing and doing it well,
and modularization of code, it is becoming obvious keeping these parts
seperate and itemize will make developing a stable whole much simpiler.
Working with the mono repo, is as should be expected, a bit nightmarish.
On the surface, it may seem like a good idea to try to keep everything
in one place, but it really is a mistake to try and do it in the source.
Al of the parts will come together when assembled, but keeping the
sources distinct will benifit the project by truly breaking things down into
the descret and components, and avoiding the need for context switching.
We can verify that each component is working as described, before
assembling them for the completed working solution, whatever that may
be. 🍌