. vuelta
Add support for vuelta, and change example code a little.

I like this btter for now, versus repeating the parsing logic, however,
this is not cleanly encapsulated, as modifying tt is not parser/plugin
specific and affects all of acorn. Will be thinking about what changes
are need to encapsulate and make specific to the plugin.

Also, right now we wipe out all the exisiting keywords, this is trivial
to avoid, and should allow a mix of the conical/english keywords, and
the localized versions, which i think will be ideal at least for
transition and in most cases, there could be a strict mode option to
only support the lcalized version and through an error otherwise.

Currentlt, we have to make changes in three places to support a keyword,
i'll be looking into why that is and if there is anyway to simplify,
howevere this will probably all be wrap up in a si.plie initilization
loop, so not a big deal, but just to understand what is going on and if
tgere are any obvious improvement wecan make that might help.

  acorn-plugin 🍀 node test/parser2.js
  {
   "type": "Program",
   "start": 0,
   "end": 68,
   "body": [
    {
     "type": "FunctionDeclaration",
     "start": 3,
     "end": 67,
     "id": {
      "type": "Identifier",
      "start": 11,
      "end": 20,
      "name": "miFunción"
     },
     "expression": false,
     "generator": false,
     "async": false,
     "params": [],
     "body": {
      "type": "BlockStatement",
      "start": 23,
      "end": 67,
      "body": [
       {
        "type": "IfStatement",
        "start": 29,
        "end": 63,
        "test": {
         "type": "Identifier",
         "start": 33,
         "end": 37,
         "name": "true"
        },
        "consequent": {
         "type": "BlockStatement",
         "start": 39,
         "end": 63,
         "body": [
          {
           "type": "ReturnStatement",
           "start": 47,
           "end": 57,
           "argument": {
            "type": "Literal",
            "start": 54,
            "end": 56,
            "value": 42,
            "raw": "42"
           }
          }
         ]
        },
        "alternate": null
       }
      ]
     }
    }
   ],
   "sourceType": "script"
  }